requirements.txt
Flask==3.1.2
psycopg2-binary==2.9.11
flask-jwt-extended==4.6.0
flask-mail==0.10.0
numpy==2.1.3
scikit-learn==1.5.2

